Implemented Reply tab with strange bug
Implemented a reply tab that dynamically shows the selected message to reply to inside a group. It's dissabled any time you click a different or same group tab on the sidenav, you close it, or  you send a message (not implemented).

TODO: There seems to be a bug where sometimes the ng-click attached to the reply button doesn't seem to be loaded in. Not sure why this is happening.