Implemented preliminary post message and reply
upon first loading, the chat bar is hidden. When you activate any chat, it pops up, and you can either write a new message or reply to an existing message. It will immediately reload the message and allow you to like or dislike or reply to it.

TODO: Make the scroll automatically land at the end, and handle excess messages?